turn on eleven more compiler checks and fix what they found
The build already refused to compile on any warning. This adds eleven more
things the compiler is now allowed to refuse on, so a class of bug that used
to reach a running node gets caught while you are still typing.

Each flag was measured before being adopted, over both configurations that
carry -Werror: the node build (1298 files, link-time optimisation on) and the
test build (2116 files, link-time optimisation off). All eleven are at zero
warnings in both, so they cost nothing today and only ever fire on new code:

  -Wundef -Wstrict-prototypes -Wdouble-promotion -Wduplicated-cond
  -Wduplicated-branches -Wshift-overflow=2 -Wattribute-alias=2 -Walloca
  -Wvla -Wtrampolines -Wflex-array-member-not-at-end

Two flags were measured and deliberately left out, with the numbers written
down next to them in the Makefile: -Wimplicit-fallthrough=5 (911 warnings at
439 sites, two of them inside the system C library's own headers) and
-Walloc-zero (0 warnings in the node build but 393 in the test build, all of
them one line re-reported through every caller). Neither is close enough to
clean to hold as a permanent rule, and weakening either to fit is not on the
table.

Four real defects came out of this:

RPC password checking leaked the password's length through timing. The
comparison stopped at the shorter of the two strings, so how long it took told
an attacker how many characters they had right. It now always looks at a fixed
512 bytes. Measured over an authentication secret growing from 16 to 496
bytes: the old code went from 2.46 ns to 13.71 ns, a 5.6x spread that tracks
the secret; the new code stays flat at about 185 ns, a 1.08x spread. Cost is
170 ns on each authentication attempt. Proven to accept and reject exactly the
same inputs as before over 3,001,156 comparisons, 938,044 of them matches,
zero disagreements.

The wallet read a transaction id through the wrong pointer type when hashing
it for its spent-output lookup. That is undefined behaviour, and the whole
node is built with optimisations that are allowed to assume it never happens,
so it was one compiler upgrade away from silently wrong lookups. Proven
identical over 20,001,548 hashes with zero disagreements, and the machine code
the compiler emits is byte-for-byte the same instruction sequence as before.

Seven files called realpath() without ever declaring it, and got away with it
only because a hardening header happened to declare it for them at the
optimisation level we ship. Building at any lower optimisation level was a
hard compile error. The whole tree now builds clean there too, down from five
errors and five warnings.

Thirteen variable-length stack arrays, sized from a value computed at run
time, are now fixed-size arrays with an explicit length check. A
variable-length array turns a weakened bounds check into a stack overrun; a
fixed array turns the same mistake into a call that simply fails. This is what
made -Wvla adoptable as a permanent rule.

Nothing under core/ was touched, no consensus predicate changed, and both
file-size gates were satisfied by re-baselining only the two files that grew
and leaving the drift count where it was.

make build-only: 0 warnings, exit 0. make lint: 119 of 119 gates pass.

+++ lib/test/src/test_mint_skip_crypto.c
@@ -1096,8 +1096,9 @@ int test_mint_skip_crypto(void)
     MSC_CHECK("fast crypto rows bind prepared epoch as checkpoint_fold",
               skip.profile_rows);
     MSC_CHECK("full fold counts all three verified stages",
-              full.script_verified == N && full.proof_verified == N &&
-              full.utxo_verified == N);
+              full.script_verified == (uint64_t)N &&
+              full.proof_verified == (uint64_t)N &&
+              full.utxo_verified == (uint64_t)N);
     MSC_CHECK("checkpoint fold never increments verified stage counters",
               skip.script_verified == 0 && skip.proof_verified == 0 &&
               skip.utxo_verified == 0);
